fix(ops): Backend.io — sm70 is cuda but takes f32 IO, not bf16
The root cause of sm70 producing degenerate output: _rows and the linear/gemm
paths cast activations to bf16 for every target.startswith('cuda'), reading
'CUDA' as 'has bf16 tensor cores'. sm70 has none — its cell is the CPU f32
kernel set (plus the fp4 GEMV) — so a bf16 activation hit an f32 kernel and
either raised (gemm_nt: 'input A dtype mismatch') or, where a cast slipped
through, collapsed the hidden state to near-uniform logits (the tiny model
raised; the 27B decoded id 220 forever).

Fix: one Backend.io attribute (bf16 only when arch=='sm90', else f32), used by
_rows; the sm70 fp4 GEMV now declares f32 X to match. Verified: tiny-model
greedy decode is now token-identical between cpu and sm70
([177,103,21,197,36,304,243,19]), where sm70 previously errored.

The fp4 GEMV parity (rel 1.5e-3) was always right — the bug was the dtype at
the op boundary, not the kernel.

+++ packages/tilerl-kernels/src/tilerl_kernels/kernels_linear.py
@@ -958,7 +958,7 @@ def make_linear_fp4_gemv_sm70(target: str):
         N, K = T.const("N, K")
         micro = 16  # 8 packed bytes = one 128-bit load; must divide `block`
         block_K = reduce_thread * micro
-        X: T.Tensor((1, K), "bfloat16")
+        X: T.Tensor((1, K), "float32")
         WQ: T.Tensor((N, K // 2), "uint8")
         Scale: T.Tensor((N, K // block), "float32")
         OScale: T.Tensor((N,), "float32")
@@ -968,7 +968,7 @@ def make_linear_fp4_gemv_sm70(target: str):
             kr = T.thread_binding(0, reduce_thread, thread="threadIdx.x")
             ni = T.thread_binding(0, n_partition, thread="threadIdx.y")
             n = bx * n_partition + ni
-            X_local = T.alloc_local((micro,), "bfloat16")
+            X_local = T.alloc_local((micro,), "float32")
             WQ_local = T.alloc_local((micro // 2,), "uint8")
             acc = T.alloc_local((1,), "float32")
             reduced = T.alloc_local((1,), "float32")
